+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - A warning message is printed when Sparv discovers source files that don't match the file extension in the corpus
   config.
 - Added `--force` flag to `run-rule` and `create-file` commands to force recreation of the listed targets.
+- Exit with an error message if a required executable is missing.
 
 ### Changed